+++ src/helpers/constants.ts
@@ -1,4 +1,5 @@
 export const tslibRoot = '/node_modules/typescript/lib';
+export const tslibPrefix = 'lib.';
 export const globalIndicator = '__global';
 export const anonymousIndicator = '__type';
 export const modulesRoot = '/node_modules/';