+++ src/helpers/identifiers.ts
@@ -14,6 +14,7 @@ import {
   OmittedExpression,
   isBindingElement,
   ArrayBindingElement,
+  ModuleName,
 } from 'typescript';
 import { typesRoot, modulesRoot, anonymousIndicator, globalIndicator } from './constants';
 
@@ -43,20 +44,21 @@ export function getTypeRefName(name: EntityName): string {
 export function getPredicateName(name: Identifier | ThisTypeNode): string {
   if (isIdentifier(name)) {
     return name.text;
-  } else {
-    // must be ThisTypeNode
+  } /* is ThisTypeNode */ else {
     return 'this';
   }
 }
 
 export function getPropName(name: PropertyName): string {
-  if (isIdentifier(name)) {
+  if (!name) {
+    return undefined;
+  } else if (isIdentifier(name)) {
     return name.text;
   } else if (isStringLiteral(name)) {
     return name.text;
   } else if (isNumericLiteral(name)) {
     return name.text;
-  } else /* isComputedPropertyName(name) */ {
+  } /* isComputedPropertyName(name) */ else {
     return name.getText();
   }
 }
@@ -74,7 +76,7 @@ export function getParameterName(name: BindingName | OmittedExpression): string 
   } else if (isArrayBindingPattern(name)) {
     const content = name.elements.map(getParameterElement).join(', ');
     return `[${content}]`;
-  } else /* is OmittedExpression */ {
+  } /* is OmittedExpression */ else {
     return '';
   }
 }
@@ -107,6 +109,11 @@ export function getLibName(fileName: string) {
   }
 
   return undefined;
+}
+
+export interface LibSpecifier {
+  packageName: string;
+  moduleName: string;
 }
 
 export function getLib(fileName: string, imports: Array<string>) {