+++ src/helpers/index.ts
@@ -3,4 +3,3 @@ export * from './node';
 export * from './paths';
 export * from './strings';
 export * from './symbol';
-export * from './type';