+++ src/output/stringify.ts
@@ -20,6 +20,9 @@ import {
   TypeMemberModel,
   TypeModelSetAccessor,
   TypeModelGetAccessor,
+  TypeModelPredicate,
+  TypeModelPrefix,
+  TypeModelPrefixReadonly,
 } from '../types';
 
 export function stringifyComment(type: WithTypeComments) {
@@ -177,6 +180,22 @@ export function stringifyGetAccessor(accessor: TypeModelGetAccessor) {
   return `${comment}${modifier}get ${accessor.name}(): ${result}`;
 }
 
+export function stringifyPredicate(predicate: TypeModelPredicate) {
+  const type = stringifyNode(predicate.value);
+  return `${predicate.name} is ${type}`;
+}
+
+export function stringifyReadonly(type: TypeModelPrefixReadonly) {
+  const value = type.value;
+
+  if (value.kind === 'ref' && value.refName === 'Array' && value.types.length === 1) {
+    const [arg] = value.types;
+    return `readonly ${stringifyNode(arg)}[]`;
+  }
+
+  return `readonly ${stringifyNode(value)}`;
+}
+
 export const enum StringifyMode {
   default = 0,
   property = 1,
@@ -201,8 +220,12 @@ export function stringifyNode(type: TypeModel, mode = StringifyMode.default) {
       return stringifyMember(type);
     case 'conditional':
       return stringifyTernary(type);
-    case 'prefix':
-      return `${type.prefix} ${stringifyNode(type.value)}`;
+    case 'readonly':
+      return stringifyReadonly(type);
+    case 'unique':
+      return `unique ${stringifyNode(type.value)}`;
+    case 'keyof':
+      return `keyof ${stringifyNode(type.value)}`;
     case 'infer':
       return `infer ${stringifyNode(type.parameter)}`;
     case 'any':
@@ -223,7 +246,7 @@ export function stringifyNode(type: TypeModel, mode = StringifyMode.default) {
     case 'unidentified':
       return 'any';
     case 'literal':
-      return JSON.stringify(type.value);
+      return `${type.value}`;
     case 'indexedAccess':
       return stringifyIndexedAccess(type);
     case 'index':
@@ -243,6 +266,10 @@ export function stringifyNode(type: TypeModel, mode = StringifyMode.default) {
       return stringifySetAccessor(type);
     case 'get':
       return stringifyGetAccessor(type);
+    case 'predicate':
+      return stringifyPredicate(type);
+    case 'parenthesis':
+      return `(${stringifyNode(type.value)})`;
   }
 
   return '';