+++ src/types/internal.ts
@@ -5,37 +5,12 @@ import * as ts from 'typescript';
  */
 declare module 'typescript' {
   interface Symbol {
-    id?: number;
     parent?: ts.Symbol;
-    target?: any;
-    declaredType?: ts.Type;
-  }
-
-  interface Type {
-    id?: number;
-    typeName?: ts.Identifier;
-    kind?: ts.SyntaxKind;
-    intrinsicName?: string;
-    parent?: ts.Type;
-    typeParameters?: Array<ts.TypeParameter>;
-    typeParameter?: ts.TypeParameter;
   }
 
   interface Node {
     symbol?: ts.Symbol;
     aliasSymbol?: ts.Symbol;
-    localSymbol?: ts.Symbol;
-  }
-
-  interface Declaration {
-    questionToken?: ts.Token<ts.SyntaxKind.QuestionToken>;
-    dotDotDotToken?: ts.Token<ts.SyntaxKind.DotDotDotToken>;
-    default?: ts.Node;
-    type?: ts.Node;
-  }
-
-  interface Expression {
-    text?: string;
   }
 
   interface SourceFile {