+++ tests/assets/enum4.ts
@@ -1,0 +1,6 @@
+enum Foo {
+  first = 1,
+  second = 2,
+}
+
+export type Bar = Foo.first;