+++ tests/assets/import1.ts
@@ -1,0 +1,3 @@
+import { EventEmitter } from 'events';
+
+export interface Foo extends EventEmitter {}