+++ tests/assets/interface4.ts
@@ -1,0 +1,5 @@
+export interface Type {
+  isUnion(): this is UnionType;
+}
+
+export interface UnionType {}